style: :art: improve structure of the backend upload handler
